jdbc/src/main/java/com/interface21/transaction/support/TransactionManager.java
package com.interface21.transaction.support;

import com.interface21.dao.DataAccessException;
import javax.sql.DataSource;
import java.sql.Connection;
import java.sql.SQLException;

public class TransactionManager {

    private final DataSource dataSource;

    public TransactionManager(DataSource dataSource) {
        this.dataSource = dataSource;
    }

    public Connection startTransaction() {
        try {
            Connection connection = dataSource.getConnection();
            connection.setAutoCommit(false);
            return connection;
        } catch (SQLException e) {
            throw new DataAccessException("Failed to start transaction", e);
        }
    }

    public void commit(Connection connection) {
        try {
            if (connection != null) {
                connection.commit();
            }
        } catch (SQLException e) {
            throw new DataAccessException("Failed to commit transaction", e);
        }
    }

    public void rollback(Connection connection) {
        try {
            if (connection != null) {
                connection.rollback();
            }
        } catch (SQLException e) {
            throw new DataAccessException("Failed to rollback transaction", e);
        }
    }
}



